requirements.txt
torch
fastapi
uvicorn[standard]
numpy 
joblib
redis
faiss-cpu
pandas
redis
scikit-learn
